requirements.txt
elevenlabs
python-dotenv
pandas
pydub
streamlit
google-generativeai
